pyproject.toml
[project]
name = "ai-crm"
version = "0.1.0"
description = "crm with ai"
authors = [
    {name = "timofeytst"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "yoyo-migrations (>=9.0.0,<10.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "passlib[bcrypt] (==1.7.4)",
    "bcrypt (==4.3.0)",
    "email-validator (>=2.0.0,<3.0.0)"
]

[tool.poetry.scripts]
api = "ai_crm.api:create_app"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.15"

[tool.ruff]
# Line length limit
line-length = 80
target-version = "py312"

[tool.ruff.lint]
# Enable all linting rules
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

[tool.ruff.lint.isort]
# Import sorting configuration
known-first-party = ["ai_crm"]
force-single-line = false
force-sort-within-sections = true

[tool.ruff.format]
# Formatting options
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
